markdown/pagemodel: Added support for Github code block
This change does not support code highlighting but it
presents the code between '```' as a code block.
This syntax is not part of the original Markdown format. But it
allows to use this Eclipse plugin to preview Github code block.

Examples of Github code block:

-----
```
function test() {
  console.log("notice the blank line before this function?");
}
```
-----

-----
```javascript
var s = "JavaScript syntax highlighting";
alert(s);
```
-----